+++ question/question.go
@@ -104,7 +104,7 @@ func New() *Question {
 func (q *Question) Prompt(str string) *Question {
 	q.prompt = str
 	q.schema.Bydefault = ""
-	q.schema.SetModifier(0)
+	q.schema.SetChecker(0)
 	return q
 }
 
@@ -114,369 +114,54 @@ func (q *Question) Default(str string) *Question {
 	return q
 }
 
-// Mod sets the modifier flags.
-func (q *Question) Mod(flag valid.Modifier) *Question {
-	q.schema.SetModifier(flag)
+// Check sets the checker flags.
+func (q *Question) Check(flag valid.Checker) *Question {
+	q.schema.SetChecker(flag)
 	return q
 }
 
 // Min sets the checking for the minimum length of a string,
-// or the minimum value of a numeric type.
-func (q *Question) Min(n interface{}) *Question {
+// or the minimum value of an integer numeric type.
+func (q *Question) Min(n int) *Question {
 	q.schema.SetMin(n)
 	return q
 }
 
 // Max sets the checking for the maximum length of a string,
-// or the maximum value of a numeric type.
-func (q *Question) Max(n interface{}) *Question {
+// or the maximum value of an integer numeric type.
+func (q *Question) Max(n int) *Question {
 	q.schema.SetMax(n)
 	return q
 }
 
 // Range sets the checking for the minimum and maximum lengths of a string,
-// or the minimum and maximum values of a numeric type.
-func (q *Question) Range(min, max interface{}) *Question {
+// or the minimum and maximum values of an integer numeric type.
+func (q *Question) Range(min, max int) *Question {
 	q.schema.SetRange(min, max)
 	return q
 }
 
-// == Read
-
-// read is the base to read and validate input.
-func (q *Question) read() (iface interface{}, err error) {
-	var hadError bool
-	line, err := q.newLine()
-	if err != nil {
-		return nil, err
-	}
-
-	for {
-		input, err := line.Read()
-		if err != nil {
-			return nil, err
-		}
-
-		switch q.schema.DataType {
-		case yoda.T_bool:
-			iface, err = valid.Bool(q.schema, input)
-		case yoda.T_int64:
-			iface, err = valid.Int64(q.schema, input)
-		case yoda.T_uint64:
-			iface, err = valid.Uint64(q.schema, input)
-		case yoda.T_float64:
-			iface, err = valid.Float64(q.schema, input)
-		case yoda.T_string:
-			iface, err = valid.String(q.schema, input)
-		default:
-			panic("unimplemented")
-		}
-
-		// Validation
-		if err == valid.ErrRequired {
-			os.Stderr.Write(readline.DelLine_CR)
-			fmt.Fprintf(os.Stderr, "%s%s", q.prefixError, err)
-			term.Output.Write(readline.CursorUp)
-			hadError = true
-			continue
-		}
-		// Error of type.
-		if err != nil {
-			os.Stderr.Write(readline.DelLine_CR)
-			fmt.Fprintf(os.Stderr, "%s%s", q.prefixError, err)
-			term.Output.Write(readline.CursorUp)
-			hadError = true
-			continue
-		}
-
-		if hadError {
-			os.Stderr.Write(readline.DelLine_CR)
-		}
-		return iface, nil
-	}
-}
-
-// ReadBool prints the prompt waiting to get a string that represents a boolean.
-func (q *Question) ReadBool() (bool, error) {
-	q.schema.DataType = yoda.T_bool
-
-	iface, err := q.read()
-	if err != nil {
-		return false, err
-	}
-	return iface.(bool), nil
-}
-
-// ReadInt64 prints the prompt waiting to get an integer number.
-func (q *Question) ReadInt64() (int64, error) {
-	q.schema.DataType = yoda.T_int64
-
-	iface, err := q.read()
-	if err != nil {
-		return 0, err
-	}
-	return iface.(int64), nil
-}
-
-// ReadUint64 prints the prompt waiting to get an unsigned integer number.
-func (q *Question) ReadUint64() (uint64, error) {
-	q.schema.DataType = yoda.T_uint64
-
-	iface, err := q.read()
-	if err != nil {
-		return 0, err
-	}
-	return iface.(uint64), nil
-}
-
-// ReadFloat64 prints the prompt waiting to get a floating-point number.
-func (q *Question) ReadFloat64() (float64, error) {
-	q.schema.DataType = yoda.T_float64
-
-	iface, err := q.read()
-	if err != nil {
-		return 0, err
-	}
-	return iface.(float64), nil
-}
-
-// ReadString prints the prompt waiting to get a string.
-func (q *Question) ReadString() (string, error) {
-	q.schema.DataType = yoda.T_string
-
-	iface, err := q.read()
-	if err != nil {
-		return "", err
-	}
-	return iface.(string), nil
-}
-
-// == Slices
-
-// ReadInt64Slice reads multiple int64.
-// You have to press Enter twice to finish.
-func (q *Question) ReadInt64Slice() (values []int64, err error) {
-	q.schema.DataType = yoda.T_Int64Slice
-
-	if _, err = q.newLine(); err != nil {
-		return nil, err
-	}
-
-	q.schema.IsSlice = true
-	term.Output.Write(readline.CRLF)
-
-	for {
-		v, err := q.ReadInt64()
-		if err != nil {
-			return nil, err
-		}
-		if v == 0 {
-			break
-		}
-		values = append(values, v)
-	}
-
-	q.schema.IsSlice = false
-	return
-}
-
-// ReadUint64Slice reads multiple uint64.
-// You have to press Enter twice to finish.
-func (q *Question) ReadUint64Slice() (values []uint64, err error) {
-	q.schema.DataType = yoda.T_Uint64Slice
-
-	if _, err = q.newLine(); err != nil {
-		return nil, err
-	}
-
-	q.schema.IsSlice = true
-	term.Output.Write(readline.CRLF)
-
-	for {
-		v, err := q.ReadUint64()
-		if err != nil {
-			return nil, err
-		}
-		if v == 0 {
-			break
-		}
-		values = append(values, v)
-	}
-
-	q.schema.IsSlice = false
-	return
-}
-
-// ReadFloat64Slice reads multiple float64.
-// You have to press Enter twice to finish.
-func (q *Question) ReadFloat64Slice() (values []float64, err error) {
-	q.schema.DataType = yoda.T_Float64Slice
-
-	if _, err = q.newLine(); err != nil {
-		return nil, err
-	}
-
-	q.schema.IsSlice = true
-	term.Output.Write(readline.CRLF)
-
-	for {
-		v, err := q.ReadFloat64()
-		if err != nil {
-			return nil, err
-		}
-		if v == 0 {
-			break
-		}
-		values = append(values, v)
-	}
-
-	q.schema.IsSlice = false
-	return
-}
-
-// ReadSliceString reads multiple strings.
-// You have to press Enter twice to finish.
-func (q *Question) ReadStringSlice() (values []string, err error) {
-	q.schema.DataType = yoda.T_StringSlice
-
-	if _, err = q.newLine(); err != nil {
-		return nil, err
-	}
-
-	q.schema.IsSlice = true
-	term.Output.Write(readline.CRLF)
-
-	for {
-		v, err := q.ReadString()
-		if err != nil {
-			return nil, err
-		}
-		if v == "" {
-			break
-		}
-		values = append(values, v)
-	}
-
-	q.schema.IsSlice = false
-	return
-}
-
-// == Choices
-
-var (
-	down2 = []byte{13, 10, 13, 10}
-	up2   = []byte(fmt.Sprintf("%s%s", readline.CursorUp, readline.CursorUp))
-)
-
-// ChoiceInt prints the prompt waiting to get an int that is in the slice.
-func (q *Question) ChoiceInt(choices []int) (int, error) {
-	term.Output.Write(down2)
-	fmt.Fprintf(term.Output, "   >>> %v", choices)
-	term.Output.Write(up2)
-
-	for {
-		value, err := q.ReadInt64()
-		if err != nil {
-			return 0, err
-		}
-		for _, v := range choices {
-			if v == int(value) {
-				return int(value), nil
-			}
-		}
-
-		os.Stderr.Write(readline.DelLine_CR)
-		fmt.Fprintf(os.Stderr, "%s%s", q.prefixError, "invalid choice")
-		term.Output.Write(readline.CursorUp)
-	}
-}
-
-// ChoiceFloat64 prints the prompt waiting to get a float64 that is in the slice.
-func (q *Question) ChoiceFloat64(choices []float64) (float64, error) {
-	term.Output.Write(down2)
-	fmt.Fprintf(term.Output, "   >>> %v", choices)
-	term.Output.Write(up2)
-
-	for {
-		value, err := q.ReadFloat64()
-		if err != nil {
-			return 0, err
-		}
-		for _, v := range choices {
-			if v == float64(value) {
-				return float64(value), nil
-			}
-		}
-
-		os.Stderr.Write(readline.DelLine_CR)
-		fmt.Fprintf(os.Stderr, "%s%s", q.prefixError, "invalid choice")
-		term.Output.Write(readline.CursorUp)
-	}
-}
-
-// ChoiceString prints the prompt waiting to get a string that is in the slice.
-func (q *Question) ChoiceString(choices []string) (string, error) {
-	term.Output.Write(down2)
-	fmt.Fprintf(term.Output, "   >>> %v", choices)
-	term.Output.Write(up2)
-
-	for {
-		value, err := q.ReadString()
-		if err != nil {
-			return "", err
-		}
-		for _, v := range choices {
-			if v == string(value) {
-				return string(value), nil
-			}
-		}
-
-		os.Stderr.Write(readline.DelLine_CR)
-		fmt.Fprintf(os.Stderr, "%s%s", q.prefixError, "invalid choice")
-		term.Output.Write(readline.CursorUp)
-	}
-}
-
-// == Regexp
-/*
-// NewRegexp sets a regular expression "re" with a name to be identified.
-func (q *Question) NewRegexp(name, re string) *Question {
-	q.val = validate.NewRegexp(q.flag, name, re)
+// SetMinFloat sets the checking for the minimum length of a floating-point
+// numeric type.
+func (q *Question) SetMinFloat(n float64) *Question {
+	q.schema.SetMinFloat(n)
 	return q
 }
 
-// SetBasicEmail sets an email based into a basic format like regular expression.
-func (q *Question) SetBasicEmail() *Question {
-	q.val = validate.NewBasicEmail(q.flag)
+// SetMaxFloat sets the checking for the maximum length of a floating-point
+// numeric type.
+func (q *Question) SetMaxFloat(n float64) *Question {
+	q.schema.SetMaxFloat(n)
 	return q
 }
 
-// SetRFCEmail sets an email based in RFC like regular expression.
-func (q *Question) SetRFCEmail() *Question {
-	q.val = validate.NewRFCEmail(q.flag)
+// SetRangeFloat sets the checking for the minimum and maximum lengths of
+// a floating-point numeric type.
+func (q *Question) SetRangeFloat(min, max float64) *Question {
+	q.schema.SetRangeFloat(min, max)
 	return q
 }
 
-// AddRegexp adds another regular expression.
-// Whether the name to identify the reg.exp. has been set, it is changed.
-func (q *Question) AddRegexp(name, re string) *Question {
-	q.val.AddRegexp(name, re)
-	return q
-}
-
-// Read prints the prompt waiting to get a string that matches with a set of
-// regular expressions.
-func (q *Question) Read() (string, error) {
-	if q.defaultVal != nil {
-		q.val.SetDefault(q.defaultVal)
-		q.suffixPrompt = msgForDefault(q.defaultVal)
-	}
-
-	value, err := q.read(q.newLine(), q.val)
-	q.clean()
-	return value.(string), err
-}
-*/
 // == Printing
 
 // The values by default are set to bold.
@@ -539,24 +224,6 @@ func (q *Question) newLine() (*readline.Line, error) {
 	return readline.NewLine(q.term, fullPrompt, q.prefixError, extraChars, nil)
 }
 
-// == Terminal handler
-
-// ExitAtCtrlD exits when it is pressed Ctrl-D, with value n.
-func (q *Question) ExitAtCtrlD(n int) {
-	go func() {
-		select {
-		case <-readline.ChanCtrlD:
-			term.Output.Write(readline.DelLine_CR)
-			os.Exit(n)
-		}
-	}()
-}
-
-// Restore restores terminal settings.
-func (q *Question) Restore() error { return q.term.Restore() }
-
-// == Utility
-
 // PrintAnswer prints values returned by a Question.
 func PrintAnswer(i interface{}, err error) {
 	term.Output.Write(readline.DelLine_CR)
@@ -574,3 +241,19 @@ func PrintAnswer(i interface{}, err error) {
 		fmt.Fprintf(term.Output, "%s\r\n", err)
 	}
 }
+
+// == Terminal handler
+
+// ExitAtCtrlD exits when it is pressed Ctrl-D, with value n.
+func (q *Question) ExitAtCtrlD(n int) {
+	go func() {
+		select {
+		case <-readline.ChanCtrlD:
+			term.Output.Write(readline.DelLine_CR)
+			os.Exit(n)
+		}
+	}()
+}
+
+// Restore restores terminal settings.
+func (q *Question) Restore() error { return q.term.Restore() }
